Fixing issues in notification module refactoring
