+++ tendrl/notifier/manager/__init__.py
