+++ tendrl/notifier/tests/base.py
@@ -1,0 +1,5 @@
+
+
+class TestCase(object):
+
+    """Test case base class for all unit tests."""
