+++ tendrl/notifier/utils/util.py
